feature(select-component): scroll up action
